requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
openai
nltk
tiktoken